requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy>=2.0.35
beautifulsoup4==4.12.2
requests==2.31.0
apscheduler==3.10.4
python-dotenv==1.0.0
html5lib==1.1
notion-client==2.2.1

